Fixed Coverity reported type issues

+++ server/src/com/cloud/network/vpn/RemoteAccessVpnManagerImpl.java
@@ -363,7 +363,7 @@ public class RemoteAccessVpnManagerImpl extends ManagerBase implements RemoteAcc
                                     for (FirewallRule vpnFwRule : vpnFwRules) {
                                         _rulesDao.remove(vpnFwRule.getId());
                                         s_logger.debug("Successfully removed firewall rule with ip id=" + vpnFwRule.getSourceIpAddressId() + " and port " +
-                                            vpnFwRule.getSourcePortStart() + " as a part of vpn cleanup");
+                                            vpnFwRule.getSourcePortStart().intValue() + " as a part of vpn cleanup");
                                     }
                                 }
                             }
